chore: prepare VIP-642 for proposal
