+++ safe-vote.json
@@ -1,0 +1,22 @@
+{
+  "version": "1.0",
+  "chainId": "56",
+  "createdAt": 1783513420362,
+  "meta": {
+    "name": "VIP-642 vote FOR",
+    "description": "",
+    "txBuilderVersion": "1.18.0",
+    "createdFromSafeAddress": "0x34221485302f6F2029660a000908B5FCABB9BC6e",
+    "createdFromOwnerAddress": "",
+    "checksum": "0xba58ff908490660f451424a97f3b9d8862c1f1b18c4aae6f4c0199a75beb3703"
+  },
+  "transactions": [
+    {
+      "to": "0x2d56dC077072B53571b8252008C60e945108c75a",
+      "value": "0",
+      "data": "0x5678138800000000000000000000000000000000000000000000000000000000000002820000000000000000000000000000000000000000000000000000000000000001",
+      "contractMethod": null,
+      "contractInputsValues": null
+    }
+  ]
+}
