+++ simulations/vip-642/arbitrumone.ts
@@ -5,10 +5,10 @@ import { initMainnetUser } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 
 import { ARBITRUM_LIQUID_STAKED_ETH } from "../../vips/vip-634/phase4Markets";
-import { AGGREGATOR, SeedCommand, buildBatch } from "../../vips/vip-647/aggregatorBatches";
-import vip647 from "../../vips/vip-647/bscmainnet";
-import { ORACLE_UPDATE } from "../../vips/vip-647/oracleFeeds";
-import { marketsToZero } from "../../vips/vip-647/zeroCollateralParams";
+import { AGGREGATOR, SeedCommand, buildBatch } from "../../vips/vip-642/aggregatorBatches";
+import vip642 from "../../vips/vip-642/bscmainnet";
+import { ORACLE_UPDATE } from "../../vips/vip-642/oracleFeeds";
+import { marketsToZero } from "../../vips/vip-642/zeroCollateralParams";
 import AGGREGATOR_ABI from "./abi/CommandsAggregator.json";
 import { checkVipEffectsPostVip, checkVipEffectsPreVip } from "./utils/checkVipEffects";
 
@@ -31,7 +31,7 @@ const encode = (cmd: SeedCommand) => ({
 
 forking(FORK_BLOCK, async () => {
   before(async () => {
-    // Seed VIP-647's Arbitrum batch in-fork, exactly as an authorized batcher would on mainnet before proposal.
+    // Seed VIP-642's Arbitrum batch in-fork, exactly as an authorized batcher would on mainnet before proposal.
     const timelock = await initMainnetUser(ARB_TIMELOCK, ethers.utils.parseEther("2"));
     const agg = new Contract(aggregator, AGGREGATOR_ABI, timelock);
     expect((await agg.batchCount()).toNumber(), "batchCount must be 0").to.equal(0);
@@ -56,7 +56,7 @@ forking(FORK_BLOCK, async () => {
 
   checkVipEffectsPreVip("arbitrumone");
 
-  testForkedNetworkVipCommands("VIP-647 Arbitrum", await vip647());
+  testForkedNetworkVipCommands("VIP-642 Arbitrum", await vip642());
 
   describe("Post-VIP behavior", () => {
     it("oracle MAIN adapters repointed to the new feeds (heartbeat-based maxStalePeriod)", async () => {
