+++ simulations/vip-642/opbnbmainnet.ts
@@ -1,7 +1,7 @@
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 
 import { OPBNB } from "../../vips/vip-634/phase4Markets";
-import vip647 from "../../vips/vip-647/bscmainnet";
+import vip642 from "../../vips/vip-642/bscmainnet";
 import { checkDeprecationPostVip, checkDeprecationPreVip } from "./utils/checkDeprecation";
 
 const FORK_BLOCK = 153977309;
@@ -9,7 +9,7 @@ const FORK_BLOCK = 153977309;
 forking(FORK_BLOCK, async () => {
   checkDeprecationPreVip([OPBNB]);
 
-  testForkedNetworkVipCommands("VIP-647 opBNB", await vip647());
+  testForkedNetworkVipCommands("VIP-642 opBNB", await vip642());
 
   checkDeprecationPostVip([OPBNB]);
 });
